+++ src/ducks/Studio/AdvancedView/Calendar.module.scss
@@ -5,10 +5,14 @@
   fill: var(--jungle-green);
 }
 
+.btn {
+  fill: var(--waterloo);
+}
+
 .icon {
   margin: -2px 0 0 10px;
 }
 
 .leftIcon {
-  margin: 0px 8px 0 0;
+  margin: 0 8px 0 0;
 }
