+++ src/pages/NewDashboards/dashboards-new/SocialTool/TopTrends/TopBar/TopBar.module.scss
@@ -1,0 +1,27 @@
+.topBar {
+  padding: 16px 24px;
+  border-bottom: 1px solid var(--porcelain);
+}
+
+.socDominance {
+  border-radius: 6px;
+  padding: 4px 12px;
+  background-color: var(--orange-light-1);
+}
+
+.sum {
+  --color: var(--orange-hover);
+  text-decoration-line: underline;
+}
+
+.tooltip {
+  z-index: 11111;
+  padding: 14px 20px;
+  width: 246px;
+  color: var(--fiord);
+}
+
+.divider {
+  width: 1px;
+  background-color: var(--porcelain);
+}
